refactor: 💡 date picker and calendar component
Authors: Robin Andersson, Fredrik Lütz
