+++ src/Molecules/DatePicker/Calendar/Calendar.tsx
@@ -1,15 +1,12 @@
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-import * as R from 'ramda';
-import { format, isSameDay, isSameMonth, isToday, isWithinInterval } from 'date-fns';
+import React, { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
+import * as R from 'ramda';
+import { isSameDay, isSameMonth, isWithinInterval } from 'date-fns';
 import { Box, Flexbox, Typography } from '../../..';
 import { getCalendar, getLocale } from '../shared/dateUtils';
-import { CalendarDayProps, Props, EdgeDay } from './Calendar.types';
-
-const FIRST_DAY = 'FIRST_DAY';
-const LAST_DAY = 'LAST_DAY';
-const NUMBER_OF_VISIBLE_WEEKS = 6;
-const NUMBER_OF_VISIBLE_DAYS = 7;
+import { Props } from './Calendar.types';
+import { FIRST_DAY, LAST_DAY, NUMBER_OF_VISIBLE_DAYS, NUMBER_OF_VISIBLE_WEEKS } from './constants';
+import { CalendarDay } from './CalendarDay';
 
 const calendarDayRefs = [...Array(NUMBER_OF_VISIBLE_WEEKS)].map(() => {
   return [...Array(NUMBER_OF_VISIBLE_DAYS)].reduce((acc) => {
@@ -17,137 +14,13 @@ const calendarDayRefs = [...Array(NUMBER_OF_VISIBLE_WEEKS)].map(() => {
   }, []);
 });
 
-const StyledWeekDay = styled(Box)`
+export const StyledBox = styled(Box)`
   border: 1px solid transparent;
   min-width: ${({ theme }) => theme.spacing.unit(10)}px;
   text-align: center;
   margin-bottom: ${({ theme }) => theme.spacing.unit(2)}px;
   margin-top: ${({ theme }) => theme.spacing.unit(3)}px;
 `;
-
-const StyledCalendarDay = styled(Box)<{
-  $disabled?: boolean;
-  $selected?: boolean;
-  $withinRange?: boolean;
-  $isToday?: boolean;
-  $edgeDay: EdgeDay | null;
-}>`
-  background: ${({ theme }) => theme.color.backgroundInput};
-  min-width: ${({ theme }) => theme.spacing.unit(10) + 2}px;
-  min-height: ${({ theme }) => theme.spacing.unit(10) + 2}px;
-  margin: ${({ theme }) => theme.spacing.unit(0.5)}px 0;
-  justify-content: center;
-  align-items: center;
-  display: flex;
-  cursor: pointer;
-  box-sizing: border-box;
-
-  ${({ $disabled, $selected, $withinRange, $isToday, $edgeDay, theme }) => `
-    ${$isToday ? `border: 1px solid ${theme.color.inputBorder};` : ''}
-    ${$withinRange ? `background: ${theme.color.datePickerWithinRangeBackground};` : ''}
-    ${
-      $withinRange && $edgeDay === FIRST_DAY
-        ? `background: linear-gradient(to right, ${theme.color.backgroundInput} 10%, ${theme.color.datePickerWithinRangeBackground});`
-        : ''
-    }
-    ${
-      $withinRange && $edgeDay === LAST_DAY
-        ? `background: linear-gradient(to left, ${theme.color.backgroundInput} 10%, ${theme.color.datePickerWithinRangeBackground});`
-        : ''
-    }
-    ${
-      $selected
-        ? `
-          background: ${theme.color.cta};
-          border: 1px solid transparent;`
-        : ''
-    }
-    ${
-      $disabled
-        ? `
-          color: ${theme.color.disabledText};
-          cursor: not-allowed;
-          border: 1px solid transparent;`
-        : ''
-    }
-  `}
-
-  &:focus {
-    outline: none;
-    border: 1px solid ${({ theme }) => theme.color.cta};
-  }
-
-  &:hover {
-    border: 1px solid ${({ theme }) => theme.color.cta};
-  }
-`;
-
-const CalendarDay = React.forwardRef<HTMLDivElement, CalendarDayProps>(
-  (
-    {
-      onFocus,
-      className = '',
-      date,
-      enabled = true,
-      disabled = false,
-      locale,
-      onClick,
-      onKeyDown,
-      sameMonth = true,
-      selected,
-      withinRange = false,
-      edgeDay,
-    },
-    ref,
-  ) => {
-    const textColor = (() => {
-      if (disabled || (typeof enabled === 'boolean' && !enabled)) {
-        return 'label';
-      }
-      if (!sameMonth && !selected) {
-        return 'label';
-      }
-      if (selected) {
-        return 'buttonText';
-      }
-      if (!selected) {
-        return 'text';
-      }
-      return '';
-    })();
-
-    const handleOnClick = useCallback(() => {
-      if (disabled) return;
-      if (onClick) onClick(date);
-    }, [date, disabled, onClick]);
-
-    const ariaLabel = format(date, 'cccc co MMMM, yyyy', {
-      locale,
-    });
-
-    return (
-      <StyledCalendarDay
-        // @ts-ignore
-        ref={ref} // TODO: ref should be allowed here, seems to be an issue with Styled Components
-        className={className}
-        $disabled={disabled || (typeof enabled === 'boolean' && !enabled)}
-        $selected={selected}
-        $withinRange={withinRange}
-        $isToday={isToday(date)}
-        $edgeDay={edgeDay}
-        onClick={handleOnClick}
-        onKeyDown={onKeyDown}
-        onFocus={() => onFocus()}
-        aria-label={ariaLabel}
-        tabIndex={0}
-      >
-        <Typography type="tertiary" color={(t) => t.color[textColor || 'text']}>
-          {format(date, 'd')}
-        </Typography>
-      </StyledCalendarDay>
-    );
-  },
-);
 
 const Calendar: React.FC<Props> = ({
   disableDate,
@@ -158,8 +31,6 @@ const Calendar: React.FC<Props> = ({
   selectedDate,
   selectedEndDate,
 }) => {
-  // const [focusDate, setFocusDate] = useState<Date>(newDate(selectedDate || viewedDate));
-
   const [[focusedWeek, focusedDay], setFocused] = useState<[number | null, number | null]>([
     null,
     null,
@@ -174,43 +45,32 @@ const Calendar: React.FC<Props> = ({
         case 'ArrowLeft':
           if (focusedDay > 0) {
             setFocused([focusedWeek, focusedDay - 1]);
-          } else if (focusedDay === 0 && focusedWeek === 0) {
-            setFocused([0, 0]);
-          } else {
+          } else if (focusedWeek !== 0) {
             setFocused([focusedWeek - 1, NUMBER_OF_VISIBLE_DAYS - 1]);
           }
-
           break;
 
         case 'ArrowRight':
           if (focusedDay < NUMBER_OF_VISIBLE_DAYS - 1) {
             setFocused([focusedWeek, focusedDay + 1]);
-          } else if (
-            focusedDay === NUMBER_OF_VISIBLE_DAYS - 1 &&
-            focusedWeek === NUMBER_OF_VISIBLE_WEEKS - 1
-          ) {
-            setFocused([NUMBER_OF_VISIBLE_WEEKS - 1, NUMBER_OF_VISIBLE_DAYS - 1]);
-          } else {
+          } else if (focusedWeek !== NUMBER_OF_VISIBLE_WEEKS - 1) {
             setFocused([focusedWeek + 1, 0]);
           }
           break;
 
         case 'ArrowUp':
-          if (focusedWeek === 0) {
-            setFocused([0, focusedDay]);
-          } else {
+          if (focusedWeek > 0) {
             setFocused([focusedWeek - 1, focusedDay]);
           }
           break;
 
         case 'ArrowDown':
-          if (focusedWeek === NUMBER_OF_VISIBLE_WEEKS - 1) {
-            setFocused([NUMBER_OF_VISIBLE_WEEKS - 1, focusedDay]);
-          } else {
+          if (focusedWeek < NUMBER_OF_VISIBLE_WEEKS - 1) {
             setFocused([focusedWeek + 1, focusedDay]);
           }
           break;
 
+        case ' ': // Spacebar
         case 'Enter':
           if (!R.isNil(focusedDateObjRef.current)) onClick(new Date(focusedDateObjRef.current));
           break;
@@ -241,9 +101,9 @@ const Calendar: React.FC<Props> = ({
       <Flexbox container aria-hidden>
         {calendar.weekDays.map((n) => (
           <Flexbox item justifyContent="center" alignItems="center" key={n}>
-            <StyledWeekDay>
+            <StyledBox>
               <Typography type="tertiary">{n}</Typography>
-            </StyledWeekDay>
+            </StyledBox>
           </Flexbox>
         ))}
       </Flexbox>
