+++ src/Molecules/DatePicker/Calendar/constants.ts
@@ -1,0 +1,4 @@
+export const FIRST_DAY = 'FIRST_DAY';
+export const LAST_DAY = 'LAST_DAY';
+export const NUMBER_OF_VISIBLE_WEEKS = 6;
+export const NUMBER_OF_VISIBLE_DAYS = 7;
